pyproject.toml
[tool.poetry]
name = "privat-bank"
version = "0.1.0"
description = ""
authors = ["Oleksandr Arshynov <oleksandrarshynov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.3"
websockets = "^12.0"
aiofiles = "^23.2.1"
aiopath = "^0.7.7"
httpx = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
